agents/data_graph_agent/requirements.txt
google-adk
python-dotenv
uvicorn
fastapi
